Make led control run as own process
Spawns a subprocess of the c program that only controls the led.
Sends a SIGINT signal to the process before sending the request
to get a temperature reading. The signal handler will turn the led
on and off. Kill <PID> or ctrl+c will terminate both the led control
and the server.